chore: move "web" into root to make GitHub Actions easier
actions/setup-node doesn't allow to indicate another folder the
npm project is in, and only looks in the root. Let's not make
life more complicated than needed, and put the npm project in
the root.

This does mean the rust project is hiding a bit more, but it will
be fine.

PS: during the move it also decided to upgrade some packages.