feat(web): completely reworked the interface
Mainly this commit replaces React with Svelte.

React is very difficult to handle in my cases, and information
propagation kept failing. As such, the frustration got the better
of me, and so we are now switching to Svelte. Not only is it a lot
faster, it is also more intuitive to use.

While doing so, reworked the interface where I found a need for it.
And as added bonus, there now is a dark-theme.