fix(industries): put information in tabs, to declutter the overview (#62)
There was a lot of information for industries presented, which made
it a bit hard to actually get an overview of what is going on. By
adding tabs, it becomes a lot more clear what is happening. It also
makes it a lot easier to work on callbacks or graphics.