fix(general): randomize the GRFID when creating a new project (#73)
Initially all GRFIDs were always "TRU1", and the user had to change
that. But the UI wasn't clear in this, as result, people don't change
it.

Instead of forcing the people to be creative, pick a random GRFID
from the Fyxxxxxx range, where y=type. This is currently unused,
so we hereby claim the range.